Move apps/nh_fwd/nh_fwd.lua to apps/lwaftr/nh_fwd.lua
