+++ src/program/snabbvmx/lwaftr/setup.lua
@@ -13,7 +13,7 @@ local ipv6_apps = require("apps.lwaftr.ipv6_apps")
 local lib = require("core.lib")
 local lwaftr = require("apps.lwaftr.lwaftr")
 local lwcounter = require("apps.lwaftr.lwcounter")
-local nh_fwd = require("apps.nh_fwd.nh_fwd")
+local nh_fwd = require("apps.lwaftr.nh_fwd")
 local pci = require("lib.hardware.pci")
 local raw = require("apps.socket.raw")
 local tap = require("apps.tap.tap")
